+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Guide: https://keepachangelog.com/en/1.0.0/
 ## Unreleased
 
 - [Details] Add `SearchEngine.retrieve(mapboxID: String, options: DetailsOptions)` function.
+- [SearchOptions] Add `SearchOptions.attributeSets` option. It allows to request of additional metadata attributes besides the basic ones in category search requests.
 
 ## 2.7.1