+++ CHANGELOG.md
@@ -6,8 +6,18 @@ Guide: https://keepachangelog.com/en/1.0.0/
 
 ## Unreleased
 
+<!-- Add changes for active work here -->
+
+## 2.8.0-beta.1
+
+- [Core] Update dependencies.
+
+**MapboxCommon**: v24.10.0-beta.2
+**MapboxCoreSearch**: v2.8.0-beta.1
+
 - [Details] Add `SearchEngine.retrieve(mapboxID: String, options: DetailsOptions)` function.
 - [SearchOptions] Add `SearchOptions.attributeSets` option. It allows to request of additional metadata attributes besides the basic ones in category search requests.
+- [SearchSuggestion] Add `SearchSuggestion.namePreferred` field to represent the preferred display name for the result.
 
 ## 2.7.1